+++ src/App.js
@@ -9,7 +9,7 @@ import PrivateRoute from './Routes/Service-Routes/PrivateRoute';
 import PublicRoute from './Routes/Service-Routes/PublicRoute';
 import { ContextProvider } from './Components/Context/Context';
 import Nav from './Components/Nav/Nav';
-import Dashboard from './Components/Dashboard/Dashboard';
+import Dashboard from './Routes/Dashboard/Dashboard';
 
 class App extends React.Component {
   render() {
